+++ AdhereR/NEWS.md
@@ -2,8 +2,10 @@
   
 ## New features
 
-  - now `AdhereR` can also be used from Julia.
-  
+## Bug fixes
+
+  - fix `CMA_polypharmacy()` crash due to not properly handling the new internal `suppress.special.argument.checks` argument (now there's also a unit test for `CMA_polypharmacy` to avoid such bugs in the future).
+
 # AdhereR 0.8.0
   
 ## New features